platform chat after final test
