+++ assets/lang/ar-SA.json
@@ -211,7 +211,8 @@
     "note":"ملاحظة",
     "article_number":"رقم المادة:",
     "avg_note":"للحصول على نتائج دقيقة للمتوسطات، يرجى تحديد البلدية والمنطقة ونوع العقار واستخدام العقار.",
-    "chat_other_response_message":"يبدو أنً السؤال غير واضح أو أن مشكلة ما قد حصلت، الرجاء المحاولة مرة أخرى."
+    "chat_other_response_message":"يبدو أنً السؤال غير واضح أو أن مشكلة ما قد حصلت، الرجاء المحاولة مرة أخرى.",
+    "large_digits_numbers_note":"تم عرض بيانات المخطط في جدول وذلك لوجود قيم ذات خانات رقمية كبيرة."
 
   
 }
