+++ lib/app/constants.dart
@@ -25,7 +25,7 @@ class Constant {
   static const String aqaratBaseUrl2 = 'https://ministry-municipality.v2202305135856227727.ultrasrv.de';
   static const String authorityChatBotBaseUrl = 'https://afnqcpcbai01.azurewebsites.net';
 
-    //static String xFunctionsAuthorityChatbotKey =  dotenv.env['X_FUNCTION_AUTHORITY_CHATBOT_KEY'] ?? 'NO_KEY_FOUND';
+    static String xFunctionsAuthorityChatbotKey =  dotenv.env['X_FUNCTION_AUTHORITY_CHATBOT_KEY'] ?? 'NO_KEY_FOUND';
     //============= NOTE FROM ZAK To FATINA:==============
     // static String xFunctionsAuthorityChatbotKey =  ;
     //1- uncomment last line:
